+++ packages/perspective/src/plan/view-action-plan.ts
@@ -452,6 +452,15 @@ export const viewActionPlan =
         });
       }
 
+      case 'actionLiquidityTournamentVote': {
+        return new ActionView({
+          actionView: {
+            case: 'actionLiquidityTournamentVote',
+            value: {},
+          },
+        });
+      }
+
       case undefined:
         return new ActionView({
           actionView: actionPlan.action,