+++ packages/services/src/test-utils.ts
@@ -34,6 +34,11 @@ export interface IndexedDbMock {
   upsertAuction?: Mock;
   hasTokenBalance?: Mock;
   saveGasPrices?: Mock;
+  saveTransactionInfo?: Mock;
+  getTransactionInfo?: Mock;
+  getBlockHeightByEpoch?: Mock;
+  saveLQTHistoricalVote?: Mock;
+  getLQTHistoricalVotes?: Mock;
 }
 
 export interface AuctionMock {
@@ -60,6 +65,11 @@ export interface MockQuerier {
   sct?: SctMock;
   shieldedPool?: ShieldedPoolMock;
   stake?: StakeMock;
+  funding?: FundingMock;
+}
+
+export interface FundingMock {
+  lqtCheckNullifier?: Mock;
 }
 
 export interface SctMock {