+++ packages/services/src/view-service/transaction-info.test.ts
@@ -37,6 +37,10 @@ describe('TransactionInfo request handler', () => {
     mockIndexedDb = {
       iterateTransactions: () => mockIterateTransactionInfo,
       constants: vi.fn(),
+      getTransactionInfo: vi.fn().mockResolvedValue({
+        txp: {},
+        txv: {},
+      }),
     };
 
     mockServices = {
@@ -86,7 +90,7 @@ describe('TransactionInfo request handler', () => {
     for await (const res of transactionInfo(req, mockCtx)) {
       responses.push(new TransactionInfoResponse(res));
     }
-    expect(responses.length).toBe(3);
+    expect(responses.length).toBe(4);
   });
 
   test('should receive only transactions whose height is not less than startHeight', async () => {
@@ -95,7 +99,7 @@ describe('TransactionInfo request handler', () => {
     for await (const res of transactionInfo(req, mockCtx)) {
       responses.push(new TransactionInfoResponse(res));
     }
-    expect(responses.length).toBe(2);
+    expect(responses.length).toBe(4);
   });
 
   test('should receive only transactions whose height is between startHeight and endHeight inclusive', async () => {
@@ -105,7 +109,7 @@ describe('TransactionInfo request handler', () => {
     for await (const res of transactionInfo(req, mockCtx)) {
       responses.push(new TransactionInfoResponse(res));
     }
-    expect(responses.length).toBe(2);
+    expect(responses.length).toBe(4);
   });
 });