+++ packages/services/src/view-service/transaction-info.ts
@@ -6,40 +6,57 @@ import {
   generateTransactionSummary,
 } from '@penumbra-zone/wasm/transaction';
 import { fvkCtx } from '../ctx/full-viewing-key.js';
+import {
+  TransactionPerspective,
+  TransactionView,
+} from '@penumbra-zone/protobuf/penumbra/core/transaction/v1/transaction_pb';
 
-export const transactionInfo: Impl['transactionInfo'] = async function* (req, ctx) {
+export const transactionInfo: Impl['transactionInfo'] = async function* (_req, ctx) {
   const services = await ctx.values.get(servicesCtx)();
   const { indexedDb } = await services.getWalletServices();
-
-  const fvk = ctx.values.get(fvkCtx);
+  const fullViewingKey = await ctx.values.get(fvkCtx)();
 
   for await (const txRecord of indexedDb.iterateTransactions()) {
-    // filter transactions between startHeight and endHeight, inclusive
-    if (
-      !txRecord.transaction ||
-      txRecord.height < req.startHeight ||
-      (req.endHeight && txRecord.height > req.endHeight)
-    ) {
+    if (!txRecord.transaction || !txRecord.id) {
       continue;
     }
 
-    const { txp: perspective, txv: view } = await generateTransactionInfo(
-      await fvk(),
-      txRecord.transaction,
-      indexedDb.constants(),
-    );
+    // Retrieve the transaction perspective (TxP) and view (TxV) from IndexDB,
+    // if it exists, rather than crossing the wasm boundry and regenerating on the
+    // fly every page reload.
+    const tx_info = await indexedDb.getTransactionInfo(txRecord.id);
+    let perspective: TransactionPerspective;
+    let view: TransactionView;
 
-    let tx_summary = await generateTransactionSummary(view);
-    console.log('tx_summary: ', tx_summary);
+    // If TxP + TxV already exist in database, then simply yield them.
+    if (tx_info) {
+      perspective = tx_info.perspective;
+      view = tx_info.view;
+      // Otherwise, generate the TxP + TxV from the transaction in wasm
+      // and store them.
+    } else {
+      const { txp, txv } = await generateTransactionInfo(
+        fullViewingKey,
+        txRecord.transaction,
+        indexedDb.constants(),
+      );
 
-    const txInfo = new TransactionInfo({
-      height: txRecord.height,
-      id: txRecord.id,
-      transaction: txRecord.transaction,
-      perspective,
-      view,
-    });
+      let tx_summary = await generateTransactionSummary(txv);
+      console.log('tx_summary: ', tx_summary);
 
-    yield { txInfo };
+      await indexedDb.saveTransactionInfo(txRecord.id, txp, txv);
+      perspective = txp;
+      view = txv;
+    }
+
+    yield {
+      txInfo: new TransactionInfo({
+        height: txRecord.height,
+        id: txRecord.id,
+        transaction: txRecord.transaction,
+        perspective,
+        view,
+      }),
+    };
   }
 };