+++ packages/bech32m/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @penumbra-zone/bech32
 
+## 14.0.0
+
+### Patch Changes
+
+- Updated dependencies [49ae3ab]
+  - @penumbra-zone/protobuf@8.0.0
+
+## 13.0.0
+
+### Patch Changes
+
+- Updated dependencies [68b8f36]
+  - @penumbra-zone/protobuf@7.2.0
+
+## 12.0.0
+
+### Patch Changes
+
+- Updated dependencies [29dd11a]
+  - @penumbra-zone/protobuf@7.1.0
+
 ## 11.0.0
 
 ### Minor Changes