+++ packages/client/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @penumbra-zone/client
 
+## 25.0.0
+
+### Patch Changes
+
+- Updated dependencies [49ae3ab]
+  - @penumbra-zone/protobuf@8.0.0
+
+## 24.0.0
+
+### Patch Changes
+
+- Updated dependencies [68b8f36]
+  - @penumbra-zone/protobuf@7.2.0
+
+## 23.0.0
+
+### Patch Changes
+
+- Updated dependencies [29dd11a]
+  - @penumbra-zone/protobuf@7.1.0
+
 ## 22.0.0
 
 ### Patch Changes