+++ packages/protobuf/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @penumbra-zone/protobuf
 
+## 8.0.0
+
+### Major Changes
+
+- 49ae3ab: LQT integration in web packages
+
+## 7.2.0
+
+### Minor Changes
+
+- 68b8f36: Update protos to the latest version
+
+## 7.1.0
+
+### Minor Changes
+
+- 29dd11a: - storage: add subaccount filter to `getOwnedPositionIds` method
+  - protobuf: sync latest changes in penumbra protobufs
+  - services: add subaccount filter to `ownedPositionIds` method in ViewService
+  - types: update indexedDB schema
+
 ## 7.0.0
 
 ### Major Changes